kafka_instance: don't delete internal topics
... because the management script shipping with kafka 0.9 forbids this.
With this change, I can now complete a full test suite run against a
kafka 0.9.0.0 cluster.  I had expected to see issues caused by the
"__consumer_offsets" topic getting left intact, but apparently that's
fine.

+++ pykafka/test/kafka_instance.py
@@ -103,7 +103,8 @@ class KafkaConnection(object):
     def flush(self):
         """Delete all topics."""
         for topic in self.list_topics():
-            self.delete_topic(topic)
+            if not topic.startswith(b'__'):  # leave internal topics alone
+                self.delete_topic(topic)
 
     def list_topics(self):
         """Use kafka-topics.sh to get topic information."""
